backend/analytics/requirements.txt
flask==2.3.2
flask-cors==4.0.0
psycopg2-binary==2.9.6
redis==4.5.5
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
scikit-learn==1.2.2
statsmodels==0.14.0
prophet==1.1.3
joblib==1.2.0
python-dotenv==1.0.0
gunicorn==20.1.0
